Move query updates back to theme
These are not shared between the two themes, so they can be moved into the new theme
